arch-arm: Extract a base class from ArmSemihosting
This will be used in a follow-up commit to add semihosting support for
RISC-V (which is very similar to Arm semihosting).
