+++ src/arch/arm/SConscript
@@ -133,7 +133,6 @@ SimObject('ArmCPU.py', sim_objects=[], tags='arm isa')
 
 DebugFlag('Arm', tags='arm isa')
 DebugFlag('ArmTme', 'Transactional Memory Extension', tags='arm isa')
-DebugFlag('Semihosting', tags='arm isa')
 DebugFlag('PMUVerbose', "Performance Monitor", tags='arm isa')
 
 # Add files generated by the ISA description.
